requirements.txt
torch
torchvision
networkx
transformers==4.30 
sentencepiece
pdf2image
fastapi[uvicorn]
uvicorn
PyPDF2
matplotlib
timm
scipy
shapely 
pyclipper
scikit-image
python-multipart 
onnxruntime 
requests
reportlab
loguru
gradio 
Pillow==8.4.0 
blobfile 
mypy 
numpy
pytest 
einops 
tensorboard 
Shapely==1.8.4 
opencv-python
opencv-python-headless==4.8.1.78
openai
googletrans==3.1.0a0
textdistance